edited json to not use updateing 8 byte integer garbage

+++ Dependencies/Include/nlohmann/json_fwd.hpp
@@ -28,8 +28,8 @@ template<template<typename U, typename V, typename... Args> class ObjectType =
          std::map,
          template<typename U, typename... Args> class ArrayType = std::vector,
          class StringType = std::string, class BooleanType = bool,
-         class NumberIntegerType = std::int64_t,
-         class NumberUnsignedType = std::uint64_t,
+         class NumberIntegerType = std::int32_t,
+         class NumberUnsignedType = std::uint32_t,
          class NumberFloatType = double,
          template<typename U> class AllocatorType = std::allocator,
          template<typename T, typename SFINAE = void> class JSONSerializer =
